refactor(BA-5008): merge _full test files into base test modules
Address review feedback: remove unnecessary _full file separation
by merging processor-level and quota tests into the base test files.
